# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: tea121-lite backend
  version: 0.1.0
paths:
...
  /api/v1/projects/{project_id}/runs:
    get:
      tags:
      - runs
      summary: List Runs
      operationId: list_runs_api_v1_projects__project_id__runs_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_RunOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - runs
      summary: Create Run
      operationId: create_run_api_v1_projects__project_id__runs_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreate'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}:
    get:
      tags:
      - runs
      summary: Get Run
      operationId: get_run_api_v1_runs__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/cancel:
    post:
      tags:
      - runs
      summary: Cancel Run
      operationId: cancel_run_api_v1_runs__run_id__cancel_post
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Page_RunOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RunOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[RunOut]
...
    RunCreate:
      properties:
        file_ids:
          items:
            type: string
          type: array
          title: File Ids
        detector_id:
          type: string
          title: Detector Id
        rule_pack_id:
          type: string
          title: Rule Pack Id
        detector_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Detector Version
        rule_pack_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Rule Pack Version
        variant:
          type: string
          title: Variant
          default: single
        mode:
          type: string
          enum:
          - normal
          - trace
          title: Mode
          default: normal
        config:
          additionalProperties: true
          type: object
          title: Config
        cwe_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cwe Id
        family:
          anyOf:
          - type: string
          - type: 'null'
          title: Family
        violation_kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Violation Kind
      type: object
      required:
      - detector_id
      - rule_pack_id
      title: RunCreate
    RunOut:
      properties:
        id:
          type: string
          title: Id
        project_id:
          type: string
          title: Project Id
        status:
          type: string
          title: Status
        detector_id:
          type: string
          title: Detector Id
        detector_version:
          type: string
          title: Detector Version
        rule_pack_id:
          type: string
          title: Rule Pack Id
        rule_pack_version:
          type: string
          title: Rule Pack Version
        analyzer_version:
          type: string
          title: Analyzer Version
        result_schema_version:
          type: string
          title: Result Schema Version
        file_ids:
          items:
            type: string
          type: array
          title: File Ids
        variant:
          type: string
          title: Variant
        mode:
          type: string
          title: Mode
        config:
          additionalProperties: true
          type: object
          title: Config
        progress:
          type: integer
          title: Progress
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        created_at:
          title: Created At
        started_at:
          title: Started At
        finished_at:
          title: Finished At
      type: object
      required:
      - id
      - project_id
      - status
      - detector_id
      - detector_version
      - rule_pack_id
      - rule_pack_version
      - analyzer_version
      - result_schema_version
      - file_ids
      - variant
      - mode
      - config
      - progress
      - error_message
      - created_at
      title: RunOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError